+++ dev/e2e_app/integration_test/permissions/permissions_many_test.dart
@@ -12,7 +12,7 @@ void main() {
 
     await _requestAndGrantCameraPermission($);
     await _requestAndGrantMicrophonePermission($);
-    await _requestAndDenyContactsPermission($);
+    await _requestAndDenyLocationPermission($);
   });
 }
 
@@ -44,17 +44,17 @@ Future<void> _requestAndGrantMicrophonePermission(
   expect($(#microphone).$(#statusText).text, 'Granted');
 }
 
-Future<void> _requestAndDenyContactsPermission(
+Future<void> _requestAndDenyLocationPermission(
   PatrolIntegrationTester $,
 ) async {
-  if (!await Permission.contacts.isGranted) {
-    expect($(#contacts).$(#statusText).text, 'Not granted');
-    await $('Request contacts permission').tap();
+  if (!await Permission.location.isGranted) {
+    expect($(#location).$(#statusText).text, 'Not granted');
+    await $('Request location permission').tap();
     if (await $.native.isPermissionDialogVisible(timeout: _timeout)) {
       await $.native.denyPermission();
       await $.pump();
     }
   }
 
-  expect($(#contacts).$(#statusText).text, 'Not granted');
+  expect($(#location).$(#statusText).text, 'Not granted');
 }